refactor(perl-lsp-rs): split text sync helper concerns (#0000)
